Made some changes to the moving foe classes.
Changed so that weaponmethods are called with meters and not pixels.
Changed so that controllers and models are removed when the body is destroyed.
Changed so the life of the foe is shown correctly.

+++ src/view/MovingFoeView.java
@@ -97,7 +97,7 @@ public class MovingFoeView {
 		
 		Vec2 lifePos = worldPos.sub(new Vec2(0,8));
 		g.setColor(Color.red);
-		float life = (this.model.getHp()/this.model.getMaxHp())*30;
+		float life = ((float)(this.model.getHp())/(float)(this.model.getMaxHp()))*30;
 		g.fillRect(lifePos.x, lifePos.y, life, 5);
 	}
 }
